external_auth.mobile: extracted schema
to separate module
